flask-servers/kasa-flask-server/docker-app/requirements.txt
python-kasa 
prometheus_client
flask
requests